Experiment with async input loop
The hope is that we can hide other processes behind this loop as well,
like updating buffers from the network: these will all live in a
`select!{}` within the `key()` async method of a wrapper KeySource (ex:
`AppKeySource`) which will also handle redraw, etc. and we can just
expose the KeySource to the active KeyMap.