Build async input loop (#1)
* Experiment with async input loop

The hope is that we can hide other processes behind this loop as well,
like updating buffers from the network: these will all live in a
`select!{}` within the `key()` async method of a wrapper KeySource (ex:
`AppKeySource`) which will also handle redraw, etc. and we can just
expose the KeySource to the active KeyMap.

* Add initial support for handling resize

* Prepare keymap abstraction and main input loop

* Restore resize handling

* Just combine Key events into UiEvents to avoid losing keypresses

* Actually propagate errors properly; document intended Keymap impl

I think the Option return type hinted that Keymaps should be able to
return without exiting the process, but it was not very clear. There's
now an explicit exit process signal attached to the AppState, and Keymap
can only return `Ok(())`, which feels slightly cleaner than using
`Option`.