Go back in time and remove Futures to simplify keymap handling
Futures were a neat idea but the ergonomics are just not there with our
use case. The Future seems to want the lifetime of the params to persist
past the context in which the Future is created—even when the future
does not actually persist that long due to `await` use.