Elide lifetimes on VimMode et al to use mode_stack; improve activation
We need to figure out a better way to manage the mode stack, however;
the current solution is untenable, since there's no way to actually
prevent the mode from being returned to the stack. We could add a flag
to VimKeymapState but that's hacky...

We almost want the stack to be a queue of IDs that point into a Map,
which can keep Modes alive; the Keymap can take ownership of the values
within the Map while using them, client code can push and pop IDs from
the stack, and the Keymap can return ownership of the Mode to the State
at the end of the loop *if* the mode's ID is still on the stack. Maybe?