Add basic prompt implementation support (#8)
* Prepare to support a "prompt" window for handling cmdline mode, etc

* Enable Buffer to measure its height

* Reorganize TUI rendering code; prepare vertical shifting for prompt

* Support visual rendering tests; Add simple Display.shift_up() test

* Add merge_at_y test + handle incomplete lines in Display::of_string

* Don't try to render prompt until it has content

* Add convenience to clear the prompt

* Actually resize the prompt as appropriate

The prompt now renders!

* Prepare "prompt mode"

The next trick will be actually wiring it up, due to Rust's lifetime
concerns...

* Elide lifetimes on VimMode et al to use mode_stack; improve activation

We need to figure out a better way to manage the mode stack, however;
the current solution is untenable, since there's no way to actually
prevent the mode from being returned to the stack. We could add a flag
to VimKeymapState but that's hacky...

We almost want the stack to be a queue of IDs that point into a Map,
which can keep Modes alive; the Keymap can take ownership of the values
within the Map while using them, client code can push and pop IDs from
the stack, and the Keymap can return ownership of the Mode to the State
at the end of the loop *if* the mode's ID is still on the stack. Maybe?

* Implement our ModeStack idea---we no longer get stuck in prompt()

* Use prompt window's cursor if focused

* Extract prompt buffer contents properly

* Implement an "after" handler

* Fix MotionContext implementation for AppState

window() was not matching current_window(), etc

* Fix starts_with check

* Don't run after handling if mode was popped