Fix cursor position and scrolling with wrapped lines (#9)
* Prepare some (currently failing) tests around wrapping and cursor pos

* Improve cursor positioning on wrapped lines

Not sure why the y_offset was subtracted, but this fixes that. We still
don't actually handle word wrapping for the cursor, however...

* Fix scrolling past full lines

Just forgot to update the scrolled_lines... woops

* Properly fix visual scroll rendering!

Now if we can just figure out proper cursor wrapping we'll be in pretty
good shape! (but that can be a separate project)

Basically my mental model of how Paragraph.scroll worked was sort of
opposite how it actually works, and the undocumented argument for
scroll() didn't help (I assumed x,y but it was line,col)

* Implement cursor wrapping in a terribly inefficient way

Basically we "paint" the chars on and after the cursor and find the
first painted char in the wrapped result, to handle whitespace elision