Scroll window to follow the cursor (#10)
* Introduce layout() step to enable mutable adjustments pre-render

If we *would* render the cursor off-screen, for example, we need to
scroll accordingly. This seems to be the cleanest way to work around
Rust's borrow checker

This commit also includes some failing tests that we would like to make
pass with this branch

* Add some initial auto-scroll efforts

* Fix cursor positioning in the presence of scroll/wrapping

* Auto-scroll backward within a wrapped line to fit cursor

* Fix following cursor "down" lines

Slight miscalculation of the "end" line. Now it seems like perhaps our
scroll_down is not quite right, since we are doing a scroll(-1), which
*should* follow through wrapped lines

* Fix off-by-one in scroll_down

Auto-scroll down in wrapped lines now works!

* Clean up an old TODO