Support command input and execution (#11)
* Add basic command registry with auto abbreviation

Includes a macro for declaring commands with familiar syntax. May or may
not try to make this automatically unwind args...

* Store commands on AppState

We have to do a bit of a dance to access them, basically taking
ownership of the handler temporarily and then returning it after
execution. It's important to be able to access them this way, though,
since we'd like to support user-created commands

* Remove old <cr> -> quit mapping

We can quit in the proper way now!