WIP: begin work to track changes to support undo/redo and repeat
Each Buffer should have its own ChangeManager which can be fed with
keys, and which should be notified when a change starts and ends, as
well as what changes happend within (so they can be recorded in the
"undo" queue of the change).

I don't know whether we'll support vim's Undo Tree; for now we'll
probably just stick with a simple, linear history.