Add initial pass for cleaner panic handling (#26)
* Add initial pass for cleaner panic handling

See #23

Uses the backtrace crate to capture the backtrace at time of panic,
and defers the handling until we've cleaned up the UI. Not perfect...
but not bad! Slightly nicer might be to clean up unhelpful lines in the
stack trace...

* Experiment with cargo cache

While builds are still quite fast, relatively, the new dependency
for generating the stack trace increased CI time by about 20s, from ~1:20
to ~1:40. Not great.

This feature is *probably* worth the extra time, but if dependency
caching can shave off some of the extra time then even better

* Fix cache action config