Add *very* rough initial python scripting support
Scripts will run in a separate thread as jobs, and talk to the
app state via message passing. Right now we build the embedded
iaido module on each load() call but since we're not actually passing
the app state into it we ought to be able to do this just once when
we initialize the runtime.

Other ideas: Jobs and ScriptingManager both communicate with app state
via message passing; we ought to be able to unify this
somewhat---probably, we'll unify around the scripting interface.