Begin rewriting command_decl as a procedural macro
This should enable us to support multiple arguments to a command (which
*looked* like it worked, but actually did not) as well as skipping the
trailing commas which were necessary due to pattern matching.

Included is some basic diagnostic output when the arg type used is not
supported yet. Still lots of work to do, like handling arg completers,
porting the rest of the arg types, etc.