Improve arg parsing diagnostics; remove unnecessary dependency
We can just use syn's Error type directly to achieve the same effect as
the crate we added!

Unsupported arg types are now spanned on the actual type, instead of on
the arg name. Various things are now simplified, and arg type parsing is
moved to its own module for cleaner organization