Support multiple command arguments via procedural macro (#53)
* Begin rewriting command_decl as a procedural macro

This should enable us to support multiple arguments to a command (which
*looked* like it worked, but actually did not) as well as skipping the
trailing commas which were necessary due to pattern matching.

Included is some basic diagnostic output when the arg type used is not
supported yet. Still lots of work to do, like handling arg completers,
porting the rest of the arg types, etc.

* Simplify arg parser definition

* Fill out remaining arg type parsers

* Improve arg parsing diagnostics; remove unnecessary dependency

We can just use syn's Error type directly to achieve the same effect as
the crate we added!

Unsupported arg types are now spanned on the actual type, instead of on
the arg name. Various things are now simplified, and arg type parsing is
moved to its own module for cleaner organization

* Update Cargo.lock to reflect removed dependency

* Handle arg completer declaration

* Clean up old command arg completer definition

* Clean up some lingering, now-unnecessary, commas