Add basic support for Python scripting (#54)
* Add *very* rough initial python scripting support

Scripts will run in a separate thread as jobs, and talk to the
app state via message passing. Right now we build the embedded
iaido module on each load() call but since we're not actually passing
the app state into it we ought to be able to do this just once when
we initialize the runtime.

Other ideas: Jobs and ScriptingManager both communicate with app state
via message passing; we ought to be able to unify this
somewhat---probably, we'll unify around the scripting interface.

* WIP: attempt to embed the script API just once

* Experiment with native module

* Fix embedded python syntax

Working now! Really need a better way to print python exceptions,
though. Our current mechanism is worthless

* Use RustPython's tools for writing exceptions

This would have been very helpful earlier....

* Pull out module struct declaration to shared code

* Reorganize python module to be somewhat easier to maintain

* Actually load the specified script file provided!

* Experiment: store fns in the scripting engine; return "refs"

Basically any fns that user code passes into iaido (from a map, event
handler, etc.) will still be held in the scripting engine, but we
generate an ID that can be used to look it up in a map for invocation

* Build a KeyHandler from ScriptingFnRef

We could probably make this a From or Into eventually. For now... we
need to figure out how to incorporate UserKeyHandlers into Keymap.
Since the Keymap lives outside the State, the unfortunate answer may
be that we just have a queue of keymap modification operations in the
Queue....

* Hackily introduce initial support for mapping keys from scripting!

* Use CommandHandlerContext for UserKeyHandler; provide keymap to next_key

This change enables join_interruptably to properly handle scripting RPC
requests when used with a CommandHandlerContext (which it always is)

* Refactor fn ref management to fix invocation

* Clean up resolved TODO

* Refactor poll_key to also process async events

This fixes join_interruptably not processing pending async events until
a key is pressed!

* More scripting init to be more rigorous and integrated

Next step: don't hardcode the script path

* Automatically detect iaido init scripts

* Fix: unable to build without python feature

* Include python stdlib in the compiled binary

* Fix warnings getting written to stderr, messing up the UI

* Improve scripting error formatting

* Add preliminary python module support

This allows users to split up config logic into multiple files, using
the built-in python import functions

* Update KeyHandlerContext to properly pass its keymap

Fixes crash using next_key() in mappings, eg for `f/F`

* Don't attempt to load init scripts that we know we don't support