Experiment with new ApiManager impl + various syntax for modular API
The goal is to be able to implement various scripting APIs *once* and
generate RPC details (payload enums) + script bindings based on that, in
a way that we don't need to keep everything in a single macro block or
even single file. Everything should be able to live in isolation,
independent from the rest, so that we can more easily maintain a
possibly-quite-large API surface.