Begin integrating apigen into the app!
To use our new IaidoCore, we generate the code that builds a python
module, creating closures that delegate to the actual fn calls. This is
a bit of a bummer, but seems to be the only way to convince RustPython
to treat it like a proper module. It is unhappy if we simply return the
IaidoCore PyObjectRef as if it were a module.