Generate RPC and language-specific bindings for scripting APIs (#56)
* Experiment with new ApiManager impl + various syntax for modular API

The goal is to be able to implement various scripting APIs *once* and
generate RPC details (payload enums) + script bindings based on that, in
a way that we don't need to keep everything in a single macro block or
even single file. Everything should be able to live in isolation,
independent from the rest, so that we can more easily maintain a
possibly-quite-large API surface.

* Begin generating RPC glue

* Stub out the Api struct + reorganize

* Don't combine property and rpc config

A method can be both a property and an RPC call, but we don't need to
overload the property attribute to do that!

* Prepare to route RPC requests to handler fns

* Add initial request routing + fix response typing

* More error checking

* Fill out request enum types to hold fn param types

* Prepare ApiManager APIs to handle new, fn-specific response payload

* Wire up RPC calls; fix various this for compat

Everything seems to be in order! Next step is integrating with the
scripting languages to generate those bindings and then, finally,
replacing the old API calls with this new system.

* Add very rough initial language binding generation support

We'll need more hooks for sure, but for now at least we have figured out
how to conditionally add language support, to match the parent crate

* Support wrapping each exposed fn based on its config

We now attach #[pyproperty] for #[property]-annotated methods!

* Implement PyValue for #[ns] structs

Probably will still need to do the make_class() thing when integrating,
but I think everything is basically ready

* Remove commented-out code structure planning

It's all done! Don't need it anymore :)

* Clean up completed TODO

* Add basic support for exposing methods to scripting

* Support RPC fns with no return value

* Add preliminary support for passing params into methods

Can't quite use it yet, however, because a fn with a String param "does
not meet the requirements on the impl of IntoPyNativeFunc"

* Restore feature flag on PythonScriptingLang impl

* Add basic support for proxying method calls to handle arg conversions

* Begin integrating apigen into the app!

To use our new IaidoCore, we generate the code that builds a python
module, creating closures that delegate to the actual fn calls. This is
a bit of a bummer, but seems to be the only way to convince RustPython
to treat it like a proper module. It is unhappy if we simply return the
IaidoCore PyObjectRef as if it were a module.

* Update python compat to use IaidoCore for echo

* Support accepting fns as method arguments!

This would be a little nicer if we could use generics, but it's not
terrible!

Now all that's left is cleaning up the old stuff, and then handling
property setters.

* Remove old scripting bindings

* Clean up naming somewhat

* Continue to simplify and reorganize

* Split up scripting namespace

* Fill out some Window APIs + fix skipping provided args in py delegate

* Implement Connection object API + basic support for returning Results

We don't actually need the result enum to be Clone (and *may* be able to
get away with cloning the request enum... but that's for another day).

* Support property setters!

* Improve error message when not using & for unsupported types