Somewhat hackily allow auto-connecting from a script
Per the comments, we must ignore ScriptManager::process() if the Api has
been removed from the context, or else we will not be able to actually
connect() properly from scripts. It seems like there ought to be a
better, safer way to handle this (perhaps we can internally take the
Receiver channel and return it before we process the message we
received? Hmm...) but this is a working stopgap for now.