Add basic search motion + support executing it against operators
Since apply_motion *requires* a VimKeymap, but CommandHandlerContext
only gets a BoxableKeymap, we have to use some `Any`-based downcast
hacks to get back to the VimKeymap. Motion cannot be boxed, so we cannot
move apply_motion to the BoxableKeymap trait either.

A real bummer, all in all, but functional---and reasonably safe, since
we know that handle_search cannot be invoked except through our
PromptConfig setup *by the VimKeymap*.