Add basic support for in-buffer search (#61)
* Scaffold out search support

See #49

* Add basic search motion + support executing it against operators

Since apply_motion *requires* a VimKeymap, but CommandHandlerContext
only gets a BoxableKeymap, we have to use some `Any`-based downcast
hacks to get back to the VimKeymap. Motion cannot be boxed, so we cannot
move apply_motion to the BoxableKeymap trait either.

A real bummer, all in all, but functional---and reasonably safe, since
we know that handle_search cannot be invoked except through our
PromptConfig setup *by the VimKeymap*.

* Fix tests

* Prepare some search tests; fix infinite loop when not found

* Support search looping around the ends of the Buffer

We also return an error if the motion had no match

* Prevent echo buffering when returning PatternNotFound error

* Cleanup/refactor code

* Support repeating the most recent search via n/N

* Sanity check empty search pattern

TODO: This should actually repeat the previous search

* Fix: giving up too early due to staying at the same unmatched char

Just had to take a step forward past the unmatched char search result

* Add limits to history size

* Fix: Forward and backward search share the same history

* Simplify code slightly

* Refactor search loop behavior

In case we ever need to disable looping or something, this is a bit of a
cleaner implementation