Wrap Connection in GameConnection, which owns a GameEngine
This will allow us to naturally keep aliases and triggers for different
connections separated, and provide a natural place to hook into sending
and receiving.

It may, however, complicate the scripting-based setup---with Judo, we
can configure all the static things like aliases and triggers before
connecting to the world. It looks like the connect() API fn will block
until connected so we *could* put the methods on the connection, or
perhaps they should be called in an "on connect" event? OR, perhaps we
add methods to the BufferApiObject, and the GameEngine is attached to
the *buffer* until the connection is established?

Will definitely need to think on how all these play with script
reloading...