Add initial framework for handling game features + basic auto completion (#76)
* Wrap Connection in GameConnection, which owns a GameEngine

This will allow us to naturally keep aliases and triggers for different
connections separated, and provide a natural place to hook into sending
and receiving.

It may, however, complicate the scripting-based setup---with Judo, we
can configure all the static things like aliases and triggers before
connecting to the world. It looks like the connect() API fn will block
until connected so we *could* put the methods on the connection, or
perhaps they should be called in an "on connect" event? OR, perhaps we
add methods to the BufferApiObject, and the GameEngine is attached to
the *buffer* until the connection is established?

Will definitely need to think on how all these play with script
reloading...

* Store GameEngine for un-connected buffers in Connections

Includes `with_buffer_engine` as an accessor for a buffer's GameEngine,
regardless of whether it is connected or not, to support configuring a
buffer's GameEngine before the connection is completed.

* Add initial support for processing incoming text before outputting

* Support using the GameEngine's Completer

* Fix: unable to get connection by input buffer ID

* Scaffold out the multiplex CompletionSource

For composing other completion sources!

* Support using MultiplexCompletionSource with Completer

* Add a test for MultiplexCompletionSource

* Implement WeightedRandomSelector

* Implement the WordIndexWeightedRandomSelector

* Add an unsafe RecencyCompletionSource

There's got to be a better way to build this....

* Install a completion source

* Support conditional processing, based on "flags"

This lets us ensure some sources only deal with text that is received or
sent and not both

* Process text sent to the mud for completion

This is convenient but we might prefer to do this at a higher level so
we can (for example) not process things that came from a mapping

* Use the WordIndexWeightedRandomSelector

* Clean up completion tokens