Fix iaido core python module init
For some reason, the old mechanism we were using would get confused if
iaido was run in its own directory; even if we commented out our own
iaido init, when running `import iaido` the interpreter would just
generate a blank module—very confusing!

This change is a *bit* sketchy since I had to peek into the
implementation to see where it was looking for its imports, but seems to
work more reliably—and is also simpler, which is nice.