WIP: Prepare support for "either"-typed fn params
RustPython has a builtin "Either" type that is meant to support
accepting a param that can be one of two different types; we're
attempting to lean on this to support setting keymaps either as string
remaps or as function invocations.

This may turn out to be a terrible idea, but right now I do prefer
having a single, overloaded fn for things like `set_keymap`, etc. At any
rate, the implementation involves some improvements to our apigen code
to both simplify things and make them more robust—so that's always nice.