Improve connection throughput by reducing pauses
We now relentlessly enqueue buffer appends to the main thread. It's
certainly not as efficient as batching all the writes in some interval,
but in practice it shouldn't be a problem. In particular, we no longer
for a write to complete before beginning the next read, and we also
don't wait for data on the wire before continuing with data we have
buffered.