Convert key loop to be fully push instead of a loop with timeout
Should keep CPU usage more consistently low due to more consistent
sleep. We basically use the Dispatcher to send a no-op fn to the main
thread as a signal that the UiEvents can be polled.

We can probably simplify further and possibly process more than a
single dispatched event per loop (as we had been doing).