Rewrite the main loop for simplicity and perf (#108)
* Add experimental Dispatcher for better running on the main thread

* Replace Job main thread dispatch with Dispatcher

* Add initial threaded connection implementation

* Add initial integration of new connection handling

* Refactor transport reading into its own struct to simplify code

* Improve connection throughput by reducing pauses

We now relentlessly enqueue buffer appends to the main thread. It's
certainly not as efficient as batching all the writes in some interval,
but in practice it shouldn't be a problem. In particular, we no longer
for a write to complete before beginning the next read, and we also
don't wait for data on the wire before continuing with data we have
buffered.

* Integrate new ConnectionRecord into Connections + GameEngine

Still need to make sure the GameConnection processes everything
correctly, and possibly rename it to GameTransport

* Ensure the GameEngine processes all incoming text *first*

* Clean up Job-specific processing

* Clean up internal state on disconnect

Fixes issue where you could not delete a connection buffer after the
remote side disconnected

* Reduce mutex contention to fix unresponsiveness while connected

We still hang when attempting to send something, however.... Not quite
sure why yet

* Add an explicit yield to try to further reduce lock contention

* fix: Not sending \r\n after the provided text

* Fix: don't drop io errors sending to a connection

* Update tests

* Remove Connection; rename ConnectionFactory -> TransportFactory

* Refactor ScriptingManager to use Dispatcher

There is now only one way to get onto the main thread!

* Minor cleanup

* Simplify process_async further

* Introduce explicit scoping to fix deadlock

* Convert key loop to be fully push instead of a loop with timeout

Should keep CPU usage more consistently low due to more consistent
sleep. We basically use the Dispatcher to send a no-op fn to the main
thread as a signal that the UiEvents can be polled.

We can probably simplify further and possibly process more than a
single dispatched event per loop (as we had been doing).

* Simplify code; improve documentation; avoid unnecessary renders

* Add "sync" tokio module

* Ensure IO errors reading TUI events aren't lost